(#79) refactor: 컨벤션에 따라 파일명 수정
