+++ src/shared/components/ui/input/InputExample.tsx
@@ -52,7 +52,7 @@ interface ExampleProps {
  *
  * @see {@link RoamReady/src/app/test/input/page.dev.tsx src/app/test/input/page.dev.tsx} - 이 컴포넌트의 사용 예시를 확인할 수 있는 테스트 페이지 경로
  */
-export default function Example({
+export default function InputExample({
   onSubmit,
   isSubmitting,
   control,
